Background colour for the scene
The ros graph plugin shows up badly in dark coloured qt themes (black ellipses on a dark background).

I usually don't like forcing colours in an themed application, but in a rendered scene, your theme doesn't have a choice over the colours of the components (black ellipses for the nodes, etc), so it's choice of matching background is often wrong. This helps with that.